模組化Timer跟Problem的功能
