+++ simple-chat/server/dst/Problem.js
@@ -1,0 +1,15 @@
+"use strict";
+exports.__esModule = true;
+var fs = require("fs");
+var ramdom_1 = require("./ramdom");
+var path = require("path");
+function updateProblem(io) {
+    fs.readFile(path.join(__dirname, './problems.json'), function (err, data) {
+        if (err)
+            throw err;
+        var problems = JSON.parse(data.toString());
+        var problem = ramdom_1.getRandomProblem(problems);
+        io.sockets.emit('updateProblem', { problem: problem });
+    });
+}
+exports.updateProblem = updateProblem;
