+++ simple-chat/server/dst/Timer.js
@@ -1,0 +1,11 @@
+"use strict";
+exports.__esModule = true;
+function updateTimer(io) {
+    var date = new Date();
+    io.sockets.emit('updateTimer', {
+        hour: date.getHours(),
+        min: date.getMinutes(),
+        sec: date.getSeconds()
+    });
+}
+exports.updateTimer = updateTimer;
