+++ simple-chat/server/dst/app.js
@@ -4,8 +4,8 @@ var express = require("express");
 var http_1 = require("http");
 var path = require("path");
 var socketIo = require("socket.io");
-var fs = require("fs");
-var ramdom_1 = require("./ramdom");
+var Problem_1 = require("./Problem");
+var Timer_1 = require("./Timer");
 var app = express();
 var http = http_1.createServer(app);
 var io = socketIo(http);
@@ -18,21 +18,10 @@ app.get('/2', function (req, res) {
 var usernames = {};
 var rooms = ['room1', 'room2', 'room3'];
 setInterval(function () {
-    var date = new Date();
-    io.sockets.emit('updateTimer', {
-        hour: date.getHours(),
-        min: date.getMinutes(),
-        sec: date.getSeconds()
-    });
+    Problem_1.updateProblem(io);
 }, 1000);
 setInterval(function () {
-    fs.readFile(path.join(__dirname, './problems.json'), function (err, data) {
-        if (err)
-            throw err;
-        var problems = JSON.parse(data.toString());
-        var problem = ramdom_1.getRandomProblem(problems);
-        io.sockets.emit('updateProblem', { problem: problem });
-    });
+    Timer_1.updateTimer(io);
 }, 3000);
 io.sockets.on('connection', function (socket) {
     socket.on('adduser', function (username) {
