+++ simple-chat/server/src/Problem.ts
@@ -1,0 +1,14 @@
+import *  as fs from 'fs';
+import {getRangeRandom , getRandomProblem } from './ramdom';
+import * as path from 'path';
+
+function updateProblem(io){
+    fs.readFile(path.join(__dirname ,'./problems.json'), (err, data) => {  
+        if (err) throw err;
+        let problems = JSON.parse(data.toString());
+        let problem = getRandomProblem(problems);
+        io.sockets.emit('updateProblem', { problem:problem});
+    }); 
+}
+
+export {updateProblem}
