+++ simple-chat/server/src/Timer.ts
@@ -1,0 +1,10 @@
+function updateTimer(io){
+    let date = new Date();
+    io.sockets.emit('updateTimer', { 
+        hour: date.getHours(),
+        min: date.getMinutes(),
+        sec: date.getSeconds(),
+    });
+}
+
+export {updateTimer}
